mob session

+++ src/main/java/com/booleanuk/Model/Entity/Enemy.java
@@ -4,7 +4,7 @@ import com.booleanuk.Model.Element.Element;
 
 public abstract class Enemy extends Entity{
 
-    public Enemy(String name, float baseHP, float baseMana, Element element, int level) {
+    public Enemy(String name, int baseHP, int baseMana, Element element, int level) {
         super(name, baseHP, baseMana, level);
         this.setElement(element);
     }
