+++ src/main/java/com/booleanuk/Model/Item/Weapon/Dagger.java
@@ -2,7 +2,7 @@ package com.booleanuk.Model.Item.Weapon;
 
 public class Dagger extends Weapon{
 
-    public Dagger(String name, int requiredLvl, float minDamage, float maxDamage, float criticalChance, float criticalMultiplier, float missChance) {
+    public Dagger(String name, int requiredLvl, int minDamage, int maxDamage, float criticalChance, float criticalMultiplier, float missChance) {
         super(name, requiredLvl, minDamage, maxDamage, criticalChance, criticalMultiplier, missChance);
     }
 }
